@@ -35,49 +35,49 @@
     "redux-saga": "0.16.0",
     "reselect": "3.0.1",
     "webpack": "4.12.0",
-    "webpack-dev-server": "3.1.4"
+    "webpack-dev-server": "3.1.14"
   },
   "devDependencies": {
     "@types/classnames": "2.2.6",
     "@types/enzyme": "3.1.15",
-    "@types/jest": "^23.3.9",
+    "@types/jest": "23.3.13",
     "@types/react": "16.7.18",
     "@types/react-dom": "16.0.11",
     "@types/react-modal": "3.1.2",
     "@types/react-redux": "6.0.2",
     "@types/react-router-dom": "4.3.1",
-    "@types/react-test-renderer": "^16.0.3",
+    "@types/react-test-renderer": "16.0.3",
     "@types/redux": "3.6.31",
-    "@types/redux-mock-store": "^1.0.0",
+    "@types/redux-mock-store": "1.0.0",
     "altinn-designsystem": "2.0.19",
     "awesome-typescript-loader": "5.0.0",
-    "babel-jest": "^23.6.0",
+    "babel-jest": "23.6.0",
     "babel-preset-react": "6.24.1",
     "cross-env": "5.1.6",
     "css-loader": "0.28.11",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.6.0",
-    "eslint": "^5.8.0",
-    "eslint-config-airbnb-base": "^12.1.0",
+    "enzyme": "3.8.0",
+    "enzyme-adapter-react-16": "1.8.0",
+    "eslint": "5.12.0",
+    "eslint-config-airbnb-base": "13.1.0",
     "html-loader": "0.5.5",
     "html-webpack-plugin": "3.2.0",
     "jest": "23.1.0",
-    "jest-environment-enzyme": "^7.0.0",
-    "jest-enzyme": "^7.0.0",
-    "jest-junit": "^5.2.0",
+    "jest-environment-enzyme": "7.0.1",
+    "jest-enzyme": "7.0.1",
+    "jest-junit": "6.1.0",
     "mini-css-extract-plugin": "0.4.0",
-    "node-sass": "^4.9.3",
-    "prettier": "1.13.5",
-    "prettier-tslint": "^0.4.0",
-    "react-test-renderer": "^16.4.0",
-    "redux-mock-store": "^1.5.3",
-    "resolve-url-loader": "^2.3.0",
+    "node-sass": "4.11.0",
+    "prettier": "1.16.1",
+    "prettier-tslint": "0.4.2",
+    "react-test-renderer": "16.7.0",
+    "redux-mock-store": "1.5.3",
+    "resolve-url-loader": "3.0.0",
     "rimraf": "2.6.2",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "7.1.0",
     "source-map-loader": "0.2.3",
-    "style-loader": "^0.21.0",
-    "svg-inline-loader": "^0.8.0",
-    "ts-jest": "^23.10.4",
+    "style-loader": "0.23.1",
+    "svg-inline-loader": "0.8.0",
+    "ts-jest": "23.10.5",
     "ts-loader": "4.4.1",
     "tslint": "5.11.0",
     "tslint-eslint-rules": "5.4.0",
