@@ -1,6 +1,5 @@
 import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
 import { mount } from 'enzyme';
-import { shallow } from 'enzyme';
 import 'jest';
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
@@ -48,14 +47,14 @@ describe('>>> shared/src/navigation/drawer/TabletDrawerMenu.tsx', () => {
       mockHandleTabletDrawerMenu = jest.fn();
     });
 
-    it('should render tablet menu and menu items must match the snapshot ', () => {
-
-      const tabletDrawerMenu = renderer.create(
+    it('should render tablet menu and menu items must match the snapshot', () => {
+      const mountWrapper = mount(
         <TabletDrawerMenu
           tabletDrawerOpen={mockTabletDrawerOpen}
           handleTabletDrawerMenu={mockHandleTabletDrawerMenu}
-        />,
+        />, { attachTo: document.getElementById('root') }
       );
+      const tabletDrawerMenu = renderer.create(mountWrapper.getElement());
       expect(tabletDrawerMenu).toMatchSnapshot();
     });
   });
@@ -69,16 +68,16 @@ describe('>>> shared/src/navigation/drawer/TabletDrawerMenu.tsx', () => {
       mockHandleTabletDrawerMenu = jest.fn();
     });
 
-    it(`should render menu button and the tablet menu must not be visible
-        and menu items must match the snapshot`, () => {
-
-        const tabletDrawerMenu = shallow(
-          <TabletDrawerMenu
-            tabletDrawerOpen={mockTabletDrawerOpen}
-            handleTabletDrawerMenu={mockHandleTabletDrawerMenu}
-          />,
-        );
-        expect(tabletDrawerMenu).toMatchSnapshot();
-      });
+    it(`should render menu button and the tablet menu must not be visible and menu items must match the snapshot`, () => {
+      const mountWrapper = mount(
+        <TabletDrawerMenu
+          tabletDrawerOpen={mockTabletDrawerOpen}
+          handleTabletDrawerMenu={mockHandleTabletDrawerMenu}
+        />, { attachTo: document.getElementById('root') }
+      );
+      // This is tested with .getDOMNode, since it won't allow us to use the react-renderer
+      // Cannot set attribute visibility of Null
+      expect(mountWrapper.getDOMNode()).toMatchSnapshot();
+    });
   });
 });
